% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Penalized.Box.test.R
\name{Penalized.Box.test}
\alias{Penalized.Box.test}
\title{Penalized.Box.test}
\usage{
Penalized.Box.test(
  x,
  lag = 10,
  type = c("Ljung-Box", "Box-Pierce", "Monti"),
  fitdf = 0,
  squared = FALSE,
  log.squared = FALSE,
  absolute = FALSE,
  weighted = TRUE,
  penalized = TRUE,
  alpha = 0.05
)
}
\arguments{
\item{x}{The residuals or initial data which is a numeric vector or univariate time series}

\item{lag}{The lag which the statistic will be based on to calculate penalized acf/pacf, default = 10}

\item{type}{the type of test, default = "Ljung-Box"}

\item{fitdf}{The number of degrees of freedom to be subtracted if x is a series of residuals, default = 0}

\item{squared}{Take the squared value of residuals to detect nonlinear processes, default=FALSE}

\item{log.squared}{Take the log of the squared residuals to detect nonlinear processes, default=FALSE}

\item{absolute}{Take the absolute value of the residuals to detect nonlinear processes, default=FALSE}

\item{weighted}{If FALSE, perform the original Box-Pierce, Ljung-Box or Monti or McLeod Li type if using a transformation, default=TRUE.}

\item{penalized}{If FALSE, perform the unpenalized tests using the sample ACF/PACF estimates, default=TRUE}

\item{alpha}{The nominal level selected, default = 0.05}
}
\value{
A list with class "htest" for penalized/traditional (Weighted) Box-Pierce,
     (Weighted) Ljung-Box, (Weighted) Monti tests
     containing statistics, parameters, and p value.
}
\description{
The Penalized (Weighted) Box-Pierce, (Weighted) Ljung-Box, (Weighted) Monti, Weighted McLeod-Li type test for fitted ARMA and detection of nonlinear Models
}
\examples{
\dontrun{
data <- arima.sim(n=100, model=list(ar=0.9))
## Fit AR(1) model
p.fit <- 1
q.fit <- 0
res <- arima(data, order=c(p.fit, 0, q.fit))$resid
out_test1 <- Penalized.Box.test(x = res, lag=10, fitdf=1, type="Ljung-Box", weighted=TRUE, penalized=TRUE)

## Fit ARMA(1,1) model
p.fit <- 1
q.fit <- 1
res <- arima(data, order=c(p.fit, 0, q.fit))$resid
out_test2 <- Penalized.Box.test(x = res, lag=10, fitdf=1, type="Ljung-Box", weighted=TRUE, penalized=TRUE)
}
}
