% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Penalized.cor.gof.R
\name{Penalized.cor.gof}
\alias{Penalized.cor.gof}
\title{Penalized.cor.gof}
\usage{
Penalized.cor.gof(
  x,
  lag = 10,
  fitdf = 0,
  type = c("LB", "BP", "M", "WLB", "WBP", "WM", "MM", "LM", "WLM", "LM.pacf",
    "WLM.pacf"),
  alpha = 0.05
)
}
\arguments{
\item{x}{A numeric vector or univariate time series}

\item{lag}{The lag which the statistic will be based on to calculate penalized acf/pacf, default = 10}

\item{fitdf}{The number of degrees of freedom to be subtracted if x is a series of residuals, default = 0}

\item{type}{The type of test which penalized acf/pacf to be used, default = "LB"}

\item{alpha}{The nominal level selected, default = 0.05}
}
\value{
An object of class penalized acf/pacf estimation for time series goodness of fit tests
}
\description{
The Penalized ACF/PACF estimator for time series goodness of fit
}
\examples{
\dontrun{
data <- arima.sim(n=100, model=list(ar=0.9))
pen_cor <- Penalized.cor.gof(x = data, lag=10, fitdf=1, type="LB", alpha=0.05)
}
}
