% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Penalized.LM.test.R
\name{Penalized.LM.test}
\alias{Penalized.LM.test}
\title{Penalized.LM.test}
\usage{
Penalized.LM.test(
  x,
  h.t,
  lag = 10,
  type = c("correlation", "partial"),
  fitdf = 0,
  weighted = TRUE,
  penalized = TRUE,
  alpha = 0.05
)
}
\arguments{
\item{x}{The residuals or initial data which is a numeric vector or univariate time series}

\item{h.t}{The sample conditional variances}

\item{lag}{The lag which the statistic will be based on to calculate penalized acf/pacf, default = 10}

\item{type}{The type of test based on acf or pacf, default="correlation"}

\item{fitdf}{The number of degrees of freedom to be subtracted if x is a series of residuals, default = 0}

\item{weighted}{-- If FALSE, perform the Li-Mak test, default=TRUE.}

\item{penalized}{If FALSE, perform the unpenalized tests using the sample ACF/PACF estimates, default=TRUE}

\item{alpha}{The nominal level selected, default = 0.05}
}
\value{
A list with class "htest" for penalized/original Li-Mak tests containing
   statistics, parameters, and p value.
}
\description{
The Penalized Li-Mak type test
}
\examples{
\dontrun{
library(fGarch)
spec <- garchSpec(model=list(ar=0.2,omega=0.2, alpha=c(0.2), beta=0))
x <- garchSim(spec, n=100)
tmp <- garchFit(~arma(1,0)+garch(1,0), data=x, trace=FALSE)
res <- residuals(tmp)
h.t <- attr(tmp, "h.t")
out_test <- Penalized.LM.test(x = res, h.t=h.t, lag=10, type="correlation", fitdf=1, penalized=TRUE)
}
}
