% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Penalized.MahdiMcLeod.test.R
\name{Penalized.MahdiMcLeod.test}
\alias{Penalized.MahdiMcLeod.test}
\title{Penalized.MahdiMcLeod.test}
\usage{
Penalized.MahdiMcLeod.test(
  x,
  lag = 10,
  fitdf = 0,
  squared = FALSE,
  log.squared = FALSE,
  absolute = FALSE,
  penalized = TRUE,
  alpha = 0.05
)
}
\arguments{
\item{x}{The residuals or initial data which is a numeric vector or univariate time series}

\item{lag}{The lag which the statistic will be based on to calculate penalized acf/pacf, default = 10}

\item{fitdf}{The number of degrees of freedom to be subtracted if x is a series of residuals, default = 0}

\item{squared}{Take the squared value of residuals to detect nonlinear processes, default=FALSE}

\item{log.squared}{Take the log of the squared residuals to detect nonlinear processes, default=FALSE}

\item{absolute}{Take the absolute value of the residuals to detect nonlinear processes, default=FALSE}

\item{penalized}{If FALSE, perform the unpenalized tests using the sample ACF/PACF estimates, default=TRUE}

\item{alpha}{The nominal level selected, default = 0.05}
}
\value{
A list with class "htest" for penalized/tranditional MahdiMcLeod test results
    containing statistics, parameters, and p value.
}
\description{
The Penalized MahdiMcLeod test for fitted ARMA and detection of nonlinear Models
}
\examples{
\dontrun{
data <- arima.sim(n=100, model=list(ar=0.9))
## Fit AR(1) model
p.fit <- 1
q.fit <- 0
res <- arima(data, order=c(p.fit, 0, q.fit))$resid
out_test1 <- Penalized.MahdiMcLeod.test(x = res, lag=10, fitdf=1, penalized=TRUE)
#'
## Fit ARMA(1,1) model
p.fit <- 1
q.fit <- 1
res <- arima(data, order=c(p.fit, 0, q.fit))$resid
out_test2 <- Penalized.MahdiMcLeod.test(x = res, lag=10, fitdf=1, penalized=TRUE)
}
}
